test: cover ode rk4 solver optimization
