+++ SPEC.md
@@ -27,7 +27,7 @@
 | **Primary Language(s)** | Python 3.11+, Rust, JavaScript, TypeScript |
 | **License**             | MIT                                        |
 | **Current Version**     | N/A                                        |
-| **Spec Version**        | 1.1.164                                    |
+| **Spec Version**        | 1.1.165                                    |
 | **Last Spec Update**    | 2026-05-14                                 |
 
 ## 2. Purpose & Mission
@@ -562,6 +562,7 @@ Active development with stable core, continuous tool expansion, and web API in p
 
 | Date       | Version | Changes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | ---------- | ------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| 2026-05-14 | 1.1.165 | Optimized the ODE solver RK4 integration loop by moving state and derivative buffers from keyed objects to indexed arrays, extracted the solver and presets into a pure module, and added Vitest coverage for analytical decay, coupled oscillator order, and solver preconditions.                                                                                                                                                                                                                         |
 | 2026-05-14 | 1.1.164 | Improved calculator bounds/value input accessibility by labeling the grouped lower-bound, upper-bound, and evaluation-point controls with a shared group name plus explicit accessible names for each field.                                                                                                                                                                                                                                                                                             |
 | 2026-05-14 | 1.1.163 | Optimized the pressure-drop calculator gas-composition hot paths by replacing repeated object-entry/value reductions with single-pass keyed loops for mixture molecular weight, total composition, and normalized composition construction.                                                                                                                                                                                                                                                               |
 | 2026-05-14 | 1.1.162 | Refactored Sidekick default tab construction into a focused helper module so `UnifiedToolsSidebar` stays below the changed-file LOC budget while preserving the runtime tab behavior introduced in 1.1.161.                                                                                                                                                                                                                                                                                             |
