+++ src/ode_solver/web/package.json
@@ -1,34 +1,42 @@
 {
-    "name":  "ode-solver",
-    "private":  true,
-    "version":  "1.0.0",
-    "type":  "module",
-    "scripts":  {
-                    "dev":  "vite",
-                    "build":  "tsc \u0026\u0026 vite build",
-                    "lint":  "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-                    "preview":  "vite preview"
-                },
-    "dependencies":  {
-                         "react":  "^18.2.0",
-                         "react-dom":  "^18.2.0",
-                         "recharts":  "^2.10.0"
-                     },
-    "devDependencies":  {
-                            "@types/react":  "^18.2.0",
-                            "@types/react-dom":  "^18.2.0",
-                            "@vitejs/plugin-react":  "^4.2.0",
-                            "autoprefixer":  "^10.4.16",
-                            "postcss":  "latest",
-                            "tailwindcss":  "^3.4.0",
-                            "typescript":  "^5.3.0",
-                            "vite":  "^7.3.2"
-                        },
-    "overrides":  {
-                      "vite":  "^7.3.2",
-                      "picomatch":  "^4.0.4",
-                      "flatted":  "^3.4.2",
-                      "lodash":  "^4.18.1",
-                      "next":  "latest"
-                  }
+    "name": "ode-solver",
+    "private": true,
+    "version": "1.0.0",
+    "type": "module",
+    "scripts": {
+        "dev": "vite",
+        "build": "tsc && vite build",
+        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+        "test": "vitest run",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "d3-array": "^3.2.4",
+        "d3-format": "^3.1.2",
+        "d3-scale": "^4.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "recharts": "^2.10.0"
+    },
+    "devDependencies": {
+        "@types/react": "^18.2.0",
+        "@types/react-dom": "^18.2.0",
+        "@vitejs/plugin-react": "^4.2.0",
+        "autoprefixer": "^10.4.16",
+        "postcss": "latest",
+        "tailwindcss": "^3.4.0",
+        "typescript": "^5.3.0",
+        "vite": "^7.3.2",
+        "vitest": "^4.0.18"
+    },
+    "overrides": {
+        "vite": "^7.3.2",
+        "picomatch": "^4.0.4",
+        "flatted": "^3.4.2",
+        "lodash": "^4.18.1",
+        "next": "latest"
+    },
+    "optionalDependencies": {
+        "@rollup/rollup-win32-x64-msvc": "^4.60.2"
+    }
 }
